+++ src/main/java/com/tc/training/cabrentals/services/CarService.java
@@ -18,6 +18,6 @@ public interface CarService {
 
   Page<Car> getAllCars( Integer pageNumber, Integer pageSize, String sortBy, Sort.Direction sortDirection, String query,
       String type, String model, String seater, String mileage, Float minPrice, Float maxPrice, Gear gear,
-      Integer tripCount, Float averageRatings, CarStatus status, FuelType fuelType, UUID centerId,
+      Integer tripCount, Double averageRatings, CarStatus status, FuelType fuelType, UUID centerId,
       final LocalDateTime startDateTime, final LocalDateTime dropDateTime );
 }
