+++ src/main/java/com/tc/training/cabrentals/services/impl/CarServiceImpl.java
@@ -42,7 +42,7 @@ public class CarServiceImpl implements CarService {
   @Override
   public Page<Car> getAllCars( Integer pageNumber, Integer pageSize, String sortBy, Sort.Direction sortDirection,
       String query, String type, String model, String seater, String mileage, Float minPrice, Float maxPrice, Gear gear,
-      Integer tripCount, Float averageRatings, CarStatus status, FuelType fuelType, UUID centerId,
+      Integer tripCount, Double averageRatings, CarStatus status, FuelType fuelType, UUID centerId,
       final LocalDateTime startDateTime, final LocalDateTime dropDateTime ) {
 
     BooleanBuilder booleanBuilder = new BooleanBuilder();
